rename theme reducer and action
