+++ store/actions/theme.js
